+++ plugins/migrations/src/components/MigrationEntityDashboard/index.ts
@@ -1,0 +1,1 @@
+export { MigrationEntityDashboard } from './MigrationEntityDashboard';