+++ plugins/migrations/src/components/MigrationResultsProvider/index.ts
@@ -1,0 +1,5 @@
+export {
+  MigrationResultsProvider,
+  useMigrationResultsContext,
+} from './MigrationResultsProvider';
+export type { MigrationResultsContextValue } from './MigrationResultsProvider';